Modify to work with grasscutter

+++ kcp-netty/src/main/java/io/jpower/kcp/netty/UkcpChannel.java
@@ -9,7 +9,7 @@ import io.netty.channel.Channel;
  */
 public interface UkcpChannel extends Channel {
 
-    int conv();
+    long conv();
 
     @Override
     UkcpChannelConfig config();
